[FLINK-19046][network] Introduce separate classes for ResultPartition types
This helps to more cleanly separate the behaviors that are specific to pipelined or blocking
shuffles, and is a preparation for adding Result Partitions that are not buffer-oriented.
